--- apps/nextjs/src/app/(main)/@navbar/page.tsx
@@ -1,5 +1,0 @@
-import Navbar from "~/app/(main)/@navbar/_components/navbar";
-
-export default function NavbarSlot() {
-  return <Navbar />;
-}
